requirements.txt
streamlit>=1.31
numpy>=1.26
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.1

# Optional (only needed if you enable Bedrock backend)
boto3>=1.34.0